requirements.txt
# general modules
django >3.2,<3.3
python-decouple
Pillow
djangorestframework
djangorestframework-simplejwt
# third party modules
markdown
django-filter
drf-yasg[validation]

# email thrid party modules
django-mail-templated
#deployment module 

#api version
djoser

#refactoring and test clean code
black
flake8